Delete unused or redundant code
