+++ bcipy/simulator/README.md
@@ -2,20 +2,28 @@
 
 ### Overview
 
-This Simulator module aims to automate experimentation by sampling EEG data from prior sessions and
-running given models
-in RSVP task loop, thus simulating a
-live session.
+This Simulator module aims to automate experimentation by sampling EEG data from prior sessions and running given models in RSVP task loop, thus simulating a live session.
 
 ### Run steps
 
-`main.py` is the entry point for program. After following, `BciPy` readme steps for setup, run the
-module from shell as
-such:
+`main.py` is the entry point for program. After following `BciPy` readme steps for setup, run the module from terminal:
 
-``python3 bcipy/simulator -d "pathToDataFolderWrapper" -p "pathToParametersFile" -sm "pathToSignalModel.pkl"``
+```
+(venv) $ python bcipy/simulator -h
+usage: simulator [-h] -d DATA_FOLDER [-g GLOB_PATTERN] -m MODEL_PATH -p PARAMETERS [-n N]
 
-Then view the appropriate output folder for LOG file(s) and metrics.
+optional arguments:
+  -h, --help            show this help message and exit
+  -d DATA_FOLDER, --data_folder DATA_FOLDER
+                        Raw data folders to be processed.
+  -g GLOB_PATTERN, --glob_pattern GLOB_PATTERN
+                        glob pattern to select a subset of data folders Ex. "*RSVP_Copy_Phrase*"
+  -m MODEL_PATH, --model_path MODEL_PATH
+                        Signal models to be used
+  -p PARAMETERS, --parameters PARAMETERS
+                        Parameter File to be used
+  -n N                  Number of times to run the simulation
+```
 
 #### Program Args
 
@@ -26,25 +34,17 @@ Then view the appropriate output folder for LOG file(s) and metrics.
   - Glob patterns can also include nested directories (ex. `"*/*Matrix_Copy*"`).
 - `p` : path to the parameters.json file used to run the simulation. These parameters will be applied to
   all raw_data files when loading.
-- `sm`: the signal model argument is necessary to generate evidence
+- `m`: all pickle (.pkl) files in this directory will be loaded as signal models.
 
 #### Sim Output Details
 
-Output folders are generally located at [simulator/generated](). 
+Output folders are generally located in the `simulator/generated` directory. Each simulation will create a new directory. The directory name will be  prefixed with `SIM` and will include the current date and time.
 
-- `result.json` captures metrics from the simulation and all the state from the simulation (like _session.json_)
-- `parameters.json` captures params used, a combination of parameters inputted by you and the `sim_parameters.json`
-- Single run Folder name: _SIM_month-day-hour:min_hash_
-- Multi Run Folder name: _SIM_MULTI_runCount_month-day-hour:min_
+- `parameters.json` captures params used for the simulation
+- `sim.log` is a log file for the simulation
 
+A directory is created for each simulation run. The directory contents are similar to the session output in a normal bcipy task. Each run directory contains:
 
-#### Configuring Parameters
-
-There are various parameters that you can tweak in `sim_parameters.json` to run simulator multiple times with the same args, specify lang_model type ...
-
-
-### Architecture
-
-![simDiagram.png](res/simDiagram.png)
-
-
+- `run_n.log` log file specific to the run
+- `session.json` session data output for the task, including evidence generated for each inquiry and overall metrics.
+- `session.xlsx` session data summarized in an excel spreadsheet with charts for easier visualization.
