--- bcipy/simulator/sim_parameters.json
@@ -1,71 +1,0 @@
-{
-  "sim_task": {
-    "value": "RSVP_COPY_PHRASE",
-    "section": "sim_config",
-    "readableName": "Simulation Task",
-    "helpTip": "Specifies Task to simulate",
-    "recommended_values": [
-      "RSVP_COPY_PHRASE",
-      "RSVP_CALIBRATION"
-    ],
-    "type": "str"
-  },
-  "sim_sample_mode": {
-    "value": "0",
-    "section": "sim_config",
-    "readableName": "Simulation Sample Mode",
-    "helpTip": "Specifies method for sampling based on sampling registry",
-    "recommended_values": [
-      "0",
-      "1"
-    ],
-    "type": "str"
-  },
-  "sim_sleep_time": {
-    "value": "0",
-    "section": "sim_config",
-    "readableName": "Simulation Time Step ",
-    "helpTip": "How long to sleep in between simulation time steps",
-    "recommended_values": [
-      "0",
-      "0.5",
-      "1"
-    ],
-    "type": "float"
-  },
-  "sim_lm_parameters": {
-    "value": "",
-    "section": "sim_config",
-    "readableName": "Path to language model parameter file ",
-    "helpTip": "",
-    "recommended_values": [
-      "/lm_parameters.json"
-    ],
-    "type": "str"
-  },
-  "sim_lm_type": {
-    "value": "UNIFORM",
-    "section": "sim_config",
-    "readableName": "Language Model Type",
-    "helpTip": "Specifies which language model to use. Default: UNIFORM",
-    "recommended_values": [
-      "UNIFORM",
-      "CAUSAL",
-      "KENLM",
-      "MIXTURE",
-      "ORACLE"
-    ],
-    "type": "str"
-  },
-  "sim_run_count": {
-    "value": "1",
-    "section": "sim_config",
-    "readableName": "How many times to run simulator",
-    "helpTip": "",
-    "recommended_values": [
-      "1",
-      "5"
-    ],
-    "type": "int"
-  }
-}
